requirements.txt
streamlit
torch
torchvision
pillow
plotly
numpy
opencv-python
